+++ bindings/go/dag/sync/conversion_test.go
@@ -1,4 +1,4 @@
-package synced
+package sync
 
 import (
 	"maps"
