+++ bindings/go/dag/sync/map_to_synced.go
@@ -1,4 +1,4 @@
-package synced
+package sync
 
 import (
 	"cmp"
