+++ bindings/go/dag/sync/synced_to_map.go
@@ -1,4 +1,4 @@
-package synced
+package sync
 
 import (
 	"cmp"
