+++ bindings/go/dag/sync/traverse_test.go
@@ -1,4 +1,4 @@
-package synced
+package sync
 
 import (
 	"context"
